==> Scripts/S2_Hypothesis_Testing.ipynb <==
import pandas as pd
import warnings
import numpy as np
from scipy import stats

# Suppress all warnings
warnings.filterwarnings('ignore')


# Load the tournament_player_stats dataframe
tournament_player_stats = pd.read_csv('tournament_player_stats.csv')
# Load the matchwise_player_stats dataframe
data = pd.read_csv('ball_to_ball_data.csv')

# Filter data for two specific teams
team1_strike_rates = tournament_player_stats[tournament_player_stats['team'] == 'RCB']['strike_rate']
team2_strike_rates = tournament_player_stats[tournament_player_stats['team'] == 'SRH']['strike_rate']

print(team1_strike_rates)
print(team2_strike_rates)

# Calculate means and standard deviations
mean_team1 = np.mean(team1_strike_rates)
mean_team2 = np.mean(team2_strike_rates)
std_team1 = np.std(team1_strike_rates, ddof=1)
std_team2 = np.std(team2_strike_rates, ddof=1)

print("Mean Strike rate of RCB:",mean_team1)
print("Mean Strike rate of SRH:",mean_team2)
print("Standard Deviation Strike rate of RCB:",std_team1)
print("Standard Deviation Strike rate of SRH:",std_team1)

alpha = 0.05 #level of significance

# Z-test two-tailed
n_team1 = len(team1_strike_rates) #sample size
n_team2 = len(team2_strike_rates)
z_score = (mean_team1 - mean_team2) / np.sqrt((std_team1**2 / n_team1) + (std_team2**2 / n_team2))
print('Z-Score:', np.abs(z_score))

p_value_z = stats.norm.sf(abs(z_score)) * 2  # Two-tailed
print('P-Value:', p_value_z)

# Critical Z-Score
z_critical = stats.norm.ppf(1-alpha/2)
print('Critical Z-Score:',z_critical)

# Two-Sample T-test
t_score, p_value_t = stats.ttest_ind(team1_strike_rates, team2_strike_rates)

alpha = 0.05 #level of significance

# Compute the degrees of freedom (df) (n_A-1)+(n_b-1)
df = len(team1_strike_rates)+len(team2_strike_rates)-2
 
# Calculate the critical t-value
# ppf is used to find the critical t-value for a two-tailed test
critical_t = stats.t.ppf(1 - alpha/2, df)

print(f'T-test: T-score = {t_score}, P-Value: {p_value_t} ,Critical-T:{critical_t}')

# ANOVA
teams = tournament_player_stats['team'].unique()
strike_rates_by_team = [tournament_player_stats[tournament_player_stats['team'] == team]['strike_rate'] for team in teams]

f_statistic, p_value_anova = stats.f_oneway(*strike_rates_by_team)

print(f'ANOVA: F-statistic = {f_statistic}, P-value = {p_value_anova}')

# Create early and late over categories
data['over_category'] = np.where(data['over'] <= 10, 'early', 'late')
print(data['over_category'])

# Calculate the number of dismissals for a specific type in early and late overs
dismissal_type = 'caught'
early_dismissals = data[(data['over_category'] == 'early') & (data['wicket_type'] == dismissal_type)].shape[0]
late_dismissals = data[(data['over_category'] == 'late') & (data['wicket_type'] == dismissal_type)].shape[0]
total_early = data[data['over_category'] == 'early'].shape[0]
total_late = data[data['over_category'] == 'late'].shape[0]

print(early_dismissals)
print(late_dismissals)
print(total_early)
print(total_late)

# Calculate proportions
p_early = early_dismissals / total_early
p_late = late_dismissals / total_late

print(p_early)
print(p_late)

# Z-test
std_pooled = np.sqrt((p_early * (1 - p_early) / total_early) + (p_late * (1 - p_late) / total_late))
z_score = (p_early - p_late) / std_pooled
p_value_z = stats.norm.sf(abs(z_score)) * 2  # Two-tailed

print(f'Z-test: Z-score = {z_score}, P-Value: {p_value_z}')

# T-test
t_score, p_value_t = stats.ttest_ind_from_stats(mean1=p_early, std1=np.sqrt(p_early * (1 - p_early) / total_early), nobs1=total_early, mean2=p_late, std2=np.sqrt(p_late * (1 - p_late) / total_late), nobs2=total_late)

print(f'T-test: T-score = {t_score}, P-value = {p_value_t}')

# Create a contingency table
contingency_table = pd.crosstab(data['wicket_type'], data['over_category'])

# Chi-Square Test
chi2_stat, p_value_chi, dof, expected = stats.chi2_contingency(contingency_table)

print(f'Chi-Square Test: Chi2 Stat = {chi2_stat}, P-value = {p_value_chi}')
